backend/requirements.txt
click==8.0.4
Faker==13.2.0
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg==3.0.9
psycopg-binary==3.0.9
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.3
